add cmk.werks.validate module
this can be used via master daily docker image, to check if werks will
be valid with master branch loading/parsing logic. It has to be tested
in master daily docker image as this is the code used to build werk db
for homepage and send mails to mailinglists.

CMK-15097

Change-Id: Ibae96a5d8635d7c46804d04e87a81836fa99f6a3