Change master branch daily builds version to [BRANCH_VERSION]-YYYY-MM-DD
This cleans up an inconsistency with the daily builds of other branches,
which already used this schema.

This change enables our version parsers to know which release a daily
build belongs to. Previously it was not possible to tell that for the
master branch builds by looking at the version number.

Change-Id: Ie4f943a98a0cfe0f2ceb16c5bbcb471e13997e87