16220 Change factory setting for "Lock user accounts after N logon failures"
The factory setting for the rule "Lock user accounts after N logon failures" changes from `unset` to `10`.

If enabled local user accounts are locked after N failed login attempts. (LDAP connected users are not affected.)

Previous to this Werk a newly created site set this value to `10` but when reset to the factory setting the option was disabled.

If you disabled this setting via "Reset to default" this setthing is now set to `10` again.

CMK-13775

Change-Id: Id1a932f98ffa64cb5247fe7c35671e75d6b820e1