Activated bandit B601 and added suppressions
This is not a vulnerability. The user that configures that rule has
access to the username and password (if the password is in the
passwordstore it must be considered to be known to the user).

The only thing the potential attacker could misuse is the IP of the
Checkmk site which could circumvent IP firewall restrictions.

CMK-11145

Change-Id: Ibc2a1aecf7c61b5499f4fc3a1e28c828dce2764e