16067 FIX Fix possible "Internal server error" while saving large formulars
CMK-14280

Core change of this commit is in cmk.gui.htmllib.html:336+

A new context manager is introduced which can remember if an input field
is related to a file-upload or not. If it is related to a file-upload,
the HTML <form enctype="..."> attribute is set to multipart/form-data,
else it is set to the DEFAULT_ENCTYPE. In this change it is still
multipart/form-update, to validate the mechanical changes first.

In a subsequent commit, the DEFAULT_ENCTYPE will be set to
"application/x-www-form-urlencoded" which will result in way smaller
HTTP bodies when sending very large forms, but this enctype
can't be used in the presence of file-uploads, thus this change.

This commit is the merged commit of all these changes.

 * Rewrite html.begin_form -> html.form_context
     * bi._config
     * cmk.gui.dashboard
     * cmk.gui.backup
     * cmk.gui.werks
     * cmk.gui.view_renderer
     * cmk.gui.prediction
     * cmk.gui.message
     * cmk.gui.login
     * cmk.gui.key_mgmt
     * cmk.gui.page_menu
     * cmk.gui.painter_options
     * cmk.gui.wato.pages.user_profile
     * cmk.gui.wato.pages.tags
     * cmk.gui.wato.pages.users
     * cmk.gui.wato.pages.user_profile.two_factor
     * cmk.gui.wato
     * cmk.gui.cce._modes
     * cmk.gui.cee
     * cmk.gui.graphing.cee
     * cmk.gui.utils.confirm_with_preview
     * cmk.gui.visuals
     * cmk.gui.views
     * cmk.gui.custom_icons
     * cmk.gui.crash_reporting
     * cmk.gui.sidebar
     * cmk.gui.cme.wato
     * cmk.gui.utils
     * cmk.gui.mobile.pages
     * cmk.gui.cee.mkp_manager
     * cmk.gui.mkeventd (1/2)
     * cmk.gui.mkeventd (2/2)
     * cmk.gui.pagetypes (1/2)
     * cmk.gui.pagetypes (2/2)
     * cmk.gui.wato.pages
     * cmk.gui.table

Change-Id: I2ae2ea4abe47dd5aa0a96e7df72e2a0e0902fd9c