agent_based v2: type parse functions
The parse functions are no longer runtime validated.
We rather are a bit more specific in the type annotations.

Note that users can now make the mistake of specifying
`string_table: StringTable`, when they should be using
`string_table: StringBytesTable`, but this is an accepted
draw back.

The reduced pain due to the dropped runtime type inspection
is worth it.

Change-Id: I7e34ba9200426848a8be0554157d1f59c75f23a3