15195 SEC Protect automation user secret against timing attacks
This Werks improves how the secret of an automation user is validated during login.
Prior to the Werk, the automation user's password was not checked in a way that is safe against (theoretical) timing attacks.
This is fixed now.

Even though this Werk improves security, it does not address an exploitable vulnerability.
To aid automated scanning we assign a CVSS score of 0.0 (None) (CVSS:3.1/AV:N/AC:L/PR:N/UI:N/S:U/C:N/I:N/A:N).

CMK-15066

Change-Id: I23aebf47b235fecd5eb83fba15384f90f8a68625