Do not update roles in the CSE edition
If the role is not in shown on the page than
`get_security_userattrs` cannot find any set roles in the POST request
and stores 0 roles, removing any rights from the user. The POST request
is send when you change the set the password of an automation user and
save.

In the CSE we deactivated the option to change automation user roles by
choice. A common error of users is to misconfigure the default
automation users. That would lead to broken sites.

Change-Id: I2ee6b8fba701b2346d4ba07951087acf353e809e