15101 FIX broken autocomplete select fields
When a configuration form page is loaded with lots of selects, it could take
a very long time for the page to load. To gain more performance, the conversion
of select-fields to more user-friendly ones would be skipped whenever the
conversion would have taken longer than 3 seconds.

This lead to the bug that in this situation, autocomplete fields would stop
working completely.

This is fixed with this werk. While regular select fields can still be skipped
to gain page-load performance, autocomplete will never be skipped now.

There are no manual changes necessary by the user for this to take effect.

Change-Id: I2b7bb3e4b05499c287265c6f29a48b12d2270c8e