16290 FIX Limit login to single session: modify sessions lookup to enable the option
Prior to this werk, the user got logged out from the UI as soon as the global
setting option 'Limit login to single session at a time' was enabled. This was
due to the fact that the active sessions lookup returned an empty result which
prevented any further notification. This werk fixes this behaviour and reinstates
the behaviour of the 2.1 version.

CMK-14613

Change-Id: I351409d937d2541f7a096c80de2bb9629276bc37