16281 FIX event console: rule changes now only apply to the customer sites that were changed
Previously when changes were made to rulepack rules for the event console,
all sites received this change and therefore activate changes had to be
applied to all sites.

This werk introduces a change whereby any modifications to rulepacks
or rulepack rules are only applied to the customer's sites that they
are configured for.

SUP-14897

Change-Id: I049bc21f5411add665871a454100388b974ddada