16205 Introduce Last seen for the REST API usage
Prior to this Werk, the last time a REST API user was used was not
displayed in WATO and "Never logged in" was displayed. This Werk
changes the behavior and now displays the date and time of when it
was last used. These are the changes introduced in this Werk:

* Last seen column shows "Never" if the user was never authenticated.
* REST API users show the date and time when they last logged in.
* In the Act. column, the login icon includes the means by which
  the authentication information was last received (bearer, cookie, etc.).

CMK-14971

Change-Id: I1fcaed88b7b03b7bbb1ece2c2268764a9b8fddc5