check-cert: Map can remove some boilerplate
In the simplest case, using `Option::map` over pattern matching
results in code that looks as if we were not dealing with an
optional parameter at all, which I think is nice.

In slightly more complex cases, such as `check_validity_not_after`,
we could use `Option::map_or` but the code is then too compact
for my taste.  So let's not overdo it.

CMK-14683

Change-Id: I68dc3ca2331c7cc5c065b314df3600b21702fab1