16045 Improved host/service label conditions in rules
In parallel to host/service label filters in visuals (see <a href="https://checkmk.com/werk/15074">Werk #15074</a>), label conditions in rule configurations are improved.
Host and service label conditions now allow for more complex configurations, using logical operators 'and', 'or', 'not'. Additionally, the labels can be grouped, and label groups can be connected through the logical operators 'and', 'or', 'not' as well. Existing label condition configurations of rules are transformed to the new data structure.

Change-Id: Ied92d3188f9d6ffd5243f36db00215fe63e4a369