16267 FIX Drop 'cmk --paths' command
This Werk removes the <tt>cmk --paths</tt> command.
It has not been proberly maintained in the last years and bitrotted over time.

From a developers point of view:

This command introduces dependencies that are the wrong way around, hence getting in the way of a more sustainable architecture.
It would be likely to further decay over time.
As unreliable information is helping nobody, we drop this command.

Building it properly instead of dropping it would require much more effort, which we believe to be disproportionate.

Change-Id: Ia60bfbb12117b1c15e993baf65f0ad72c1037e65