16034 Visualization of rule search
The visualization for the rule search has been updated to provide
clearer indications of the displayed information.

The previous green overlay highlighting has been replaced by an
icon-based visualization.
Ineffective rules are denoted with a negative icon whereas
all other rule searches are marked with a green checkmark icon.

It is now clearly separated whether you are matching for a host
or searching for rules via the rule search.

CMK-14016

Change-Id: Ifa8fc4621701c45d9dea8c7203612d3dddb52619