16268 FIX Date macro replacement in fileinfo patterns
The grouping patterns in the ruleset <i>Group patterns</i> allow to specify certain date macros (for details refer to the rulesets inline help).
Previously only one instance of the macros could be used. Now all occurrences of either macro type are replaced.
In addition, the pattern no longer has to begin with a "/" or a letter (which was not documented).

Change-Id: I2ec114ace041249f85f8c8065a9418305dde16eb