Licensing: Distribute license information to remote CME sites
Currently, license verification for a CME is not required and therefore the information required for verification is not distributed across the sites in a CME setup.
With the 2.3, the CME will gain the CCE feature set, the verification will be required from then on.
To avoid license warnings on the remote sites during the update, with this werk the necessary information will be available in the whole setup after the license usage has been submitted.
If the associated subscription is valid for a 2.3 CME, no additional licensing is required during or after the update.

CMK-15165

Change-Id: Ie7bb33caec4c0ea3d46a1e8bebc57cc65568d940