16348 FIX Support Diagnostics: Use pip3
When creating a Support Diagnowtics dump, the information about the installed Python modules was collected using pip.
This actually collected information about the OS' Python environment, not about the one in the site context.
Sometimes, it caused a crash, when there was no pip installed.

Now, pip3 is used, which is built-in to the Checkmk Site.

Change-Id: I8ae4df61da24bc45fa143929825b5cff0d075244