plugins test: assert on list of failed check names
So far, the test was asserting on a boolean result of the
function process_check_result(), which made the AssertionFailed
exception output pretty meaningless. Although that is not an issue
for a local CLI execution, it is needlessly hiding the failure
for CI runs. The process_check_result() function will now return
a list of failed checks, which is expected to be empty and will
be returned as part of the AssertionFailed exception otherwise.

NOTE: The detailed diff will still be logged as before.
Change-Id: I8f5ebdd28dba2727544b5d50ced86a38b9e700ba