plugins_integration.confest: Fix site-teardown issue
Two session-scoped fixtures are initializing and tearing down two
test-sites with the same site-name. This is causing an error since the
second teardown process is attempting to remove a site that was already
removed.
Here we rename one of those site-name to avoid such conflict.
Moreover, we re-enable the logs-saving process during teardown.

Change-Id: I89b295c24db2ea4fc8c6eaa64db51a04939839bd