16163 SEC jar_signature: Prevent priviledge escalation
Agent plugin was vulnerable to the privilege escalation to root from an oracle user.

A malicious oracle user could replace the jarsigner binary with another script and put
it in the JAVA_HOME directory. The script would be executed by the root user.

The jarsigner is now executed by the oracle user, preventing the privilege escalation.

This issue was found during internal review.

CMK-15315

Change-Id: I2954f72ad3f426b10deadf4028926996a5569964