16358 Session termination after maximum session age
To mitigate the risk of a session takeover you can now define a maximum session duration.
This is set by default to 24 hours.
For the case that somebody is actively using the session over this duration you can also configure a threshold to warn users about the upcoming session termination.
This value is set by default to 15 minutes.

You can disable this feature completely via the global setting *Session management*.

Change-Id: Ic72e62c6dfd7cc29f7194d85d3cc877449956f1e