test_plugins: Force dump prefix in filename
Injected dump filename must now start either with "snmp-" or
"agent-X.X.XpX".

This should make clearer what kind of file we are injecting in the
test-site and which agent version is used to retrieve the dump.

The qa-test-data module commit-id is also updated, so that to pull the
related changes in the dumps' and responses' names.

CMK-15273

Change-Id: I851de357732d5ce5ffe31c3ef4e6d404168ac74c