adds the interface name to the inventory
For now there is only the interface index, alias and description in the inventory.
This PR will add the interface name for snmp based devices.
The name is on some devices more usefull than the description or the alias and it
will also help with the creation of the data file needed for the Network visualization add on.