werk collect: fix wrong version information for mixed werks
before this change markdown werks and legacy werks were not merged
correctly in collect subcommand.

merging was based on filename (which is different for legacy and
markdown werks) so markdown werks would overwrite the previously written
legacy werks.
now the werk id is used for merging instead of the werk filename.

Change-Id: I5d76aba7382351758dc422063de799c22a39cc0e