15316 FIX solaris: uptime: Cannot render negative timespan
Solaris server may keep a bad time when moved to different hardware.
This is corrected some minutes after booting via NTP server.

If the checkmk client is used before the time has been corrected, it was
possible that a negative time span was reported for the time since the last
start. This caused a crash in the check.

The new behaviour is, that the check will report an UNKNOWN status alongside the
message: "Your Solaris system reported to be booted in the future."

SUP-16803

Change-Id: I7db17a3330da781467544839e7c4b6c1b0bbc86a