Cleanup dead cmk-update-agent build targets
In the CI build jobs, these artifacts are built by dedicated jobs and
loaded into the package build job by early build stages.

The make logic we remove now was kind of a fallback mechanism which
might have worked one day locally. But it was broken in multiple places.
To unconfuse everything, we simply drop that case and fail the build in
case the expected artifacts are missing instead of trying to build them
ad-hoc.

Change-Id: I71e3136e9ca36c1a139b878942c7267252771157