16286 FIX bi: new endpoint to get bi aggregation state via GET method
This werk introduces a new endpoint which replicates the following endpoint

POST - /domain-types/bi_aggregation/actions/aggregation_state/invoke

We now provide the same behaviour via a GET endpoint.

GET - /domain-types/bi_aggregation/actions/aggregation_state/invoke

This means that the parameters "filter_names" and "filter_groups" which
are passed as request parameters for the POST endpoint, have to be passed
as query parameters for this new GET endpoint.

CMK-14623

Change-Id: I57008a48b12b41b0959609f59b02a25446d30a6e