Completely decouple ParsedQuery from Table.
Passing a handful of functions is a bit ugly, but at least it
makes the dependencies explicit and minimal.  Some stuff should
probably be moved out of ParsedQuery, anyway.

In a perfect world, ParsedQuery would need none of those functions,
it would just do what it name says: Parsing.  Perhaps we can slowly
move it into that direction in the future, this would make things
much cleaner and more easily testable.

Change-Id: I5f126a64c9e08558b23805fb6677c416ca7f8275