use v2 rendering III
The old rendering allowed negative values.
We replace it either by `render.time_offset`, which behaves the
same way, or by `render.timespan`, which raises a ValueError if
a negative time is encountered.
Using `timespan` in the wrong places means the plugin crashes
when it observes a negative time, while using `time_offset`
in the wrong places means we mask other problems (if the time
actually never should be negative).
We're using our best judgement here, but we might be wrong.

Change-Id: I1e20f170d3eacd5382a0baa422df310978cbe5e2