run werk related scripts in gerrit change validation
those commands are only run during release or with cronjobs, so we want
to detect problems as early as possible.

also add an option to the collect subcommand to be able to use HEAD
instead of autodiscovering branch refs. There are some branches
available in CI, but they are not automatically fetched, so they still
have invalid html and other stuff.

Change-Id: Ic8c220c5a2e6f565273214370703c275a1219f84