Fix logging of password hash failures
During the update_actions when checking if the password hash of a
automation user got out of sync exceptions can be raised if the hash is
too old.

For "normal" users we updated the hashes a while back but since we did
not use the automation user hashes we kept them as they were.
Now we want to keep the synchronous so we need to check them. In a
update action we therefore compare them to the automation.secret
content. The verify function logs these exceptions rightfully but that
clutters the output of the update action.

Since the update action uses the new `matches` function we suppress this
exception here. That matches function is only used in the update action
and during the login of an automation user so we can pass on the
logging.

Change-Id: Ib5feb8ec487c02239b43cd5f66a5c9f1eca85468