Remove unittest for "/32" nets/ips
The function is called `_normalize_ip_addresses` but it does not do
anything with the `x.x.x.x/32`. This `/32` notation (not sure if it is
really an ip address rather than a ip net) is not really normalized nor
is it used anywhere else.

So I drop the test for now so in a later change it will not raise issues
since the `/` is usually forbidden in hostnames and ip addresses...

Change-Id: I20ec012a2f217c2d189be2cc8c147c097dce089c