f12 without sudo password (1/N)
In order to comply with our internal rules we need to implement more
strict rules for sudoers configuration. This change tries to minimize
the number of times you have to authenticate with your sudo password
when executing f12.

It accomplishes this by:
* applying ACLs to the files owned by root, so your normal development
  user can modify them without additional authentication
* allowing your normal development user to execute arbitrary commands as
  site user

Applying the ACLs allows us to remove a lot of sudo invocations from
commands that copy files into the version folder. `omd-vonheute` is
also adapted to set the correct ACL so your development account can
write those files without the need of root permissions. This is done
with `omd-setup-version-for-dev` from zeug_cmk/bin

To allow the execution of arbitrary commands as site user, we need to
modify the sudoers configuration. This can be done with
`omd-setup-site-for-dev` from zeug_cmk/bin, which is also automatically
executed in `omd-vonheute`.

You have to input your sudo password for both commands, but as they are
executed in the context of `omd-vonheute` you already have authenticated
for sudo.

If you want to f12 into a site that has been created before the changes
in `omd-vonheute` you can execute the commands manually:

```
omd-setup-version-for-dev site $SITE_NAME
omd-setup-site-for-dev
```

= NOTES

The following changes were made to the .f12 scripts:

== rsync
`-a` is `-rlptgoD` but we condense it to `-rlD` because
the following can only be executed as root user:
```
-t, --times                 preserve modification times
-g, --group                 preserve group
-o, --owner                 preserve owner (super-user only)
-p, --perms                 preserve permissions
```

== PIP
--target was added, otherwise pip would silently install into the .local
  directory of the current user, which could lead to confusion when
  checkmk was not installed with `omd-vonheute` or ACLs are not correctly
  applied
--isolated was added, so user configuration can not interfere

= RESULT

== working
./packages/cmk-mkp-tool/.f12
./packages/cmc/test/.f12
./packages/cmk-rulesets/.f12
./packages/cmk-graphing/.f12
./packages/cmk-server-side-calls/.f12
./packages/frontend_vue/.f12
./packages/cmk-agent-receiver/.f12
./packages/check-cert/.f12
./packages/cmk-agent-based/.f12
./packages/cmk-werks/.f12
./packages/check-http/.f12
./packages/cmk-livestatus-client/.f12
./packages/livestatus/.f12
./packages/livestatus/test/.f12
./packages/unixcat/.f12
./non-free/cmc-protocols/.f12
./agents/special/.f12
./agents/plugins/.f12
./agents/.f12
./agents/windows/plugins/.f12
./web/htdocs/openapi/.f12
./notifications/.f12
./cmk/gui/wsgi/.f12
./cmk/gui/openapi/.f12
./cmk/gui/.f12
./cmk/.f12
./cmk/ec/.f12
./cmk/base/.f12
./cmk/cee/robotmk/.f12
./active_checks/.f12
./doc/.f12
./doc/plugin-api/.f12
./doc/documentation/.f12
./locale/.f12

== touching files outside of versions
./packages/cmc/.f12
./non-free/cmk-update-agent/.f12
./doc/predictive/.f12
./web/.f12
./doc/treasures/checks/.f12
./omd/packages/omd/.f12
./omd/packages/apache-omd/.f12
./omd/packages/maintenance/.f12
./omd/packages/check_mk/.f12
./omd/packages/check_mk/skel/etc/init.d/.f12
./omd/packages/mk-livestatus/.f12
./omd/packages/nagios/.f12
./omd/packages/mod_wsgi/.f12
./omd/packages/enterprise/alert_handlers/.f12
./omd/packages/enterprise/bin/.f12
./omd/packages/msitools/.f12

== broken
./packages/neb/test/.f12
./packages/neb/.f12
  Building rrdtool library...
  make: *** No rule to make target 'rrdtool-build-library'.  Stop.

== need root
./bin/.f12

Change-Id: I642213fcb878910025e8a7faa5d5ae0dcce0efe7