Fix security logging path
Not exactly sure why it works currently... But when I wanted to use this
in omdlib it failed, because var/log/var/log/security.log did not exist.
And the failure is more reasonable.

The problem though was the mocking. In the unittests we apparently set
the paths to a temppath in `patch_omd_site` but that happens after we
had the import of log_dir in the `_log.py` file so that variable was
never mocked... That cost me more time than I want to admit so I add
this notice here...

Change-Id: I6c897e628074cb87ea5aa779058135ca90e4d534