15721 FIX Fix wrong edition reported via livestatus in enterprise sites running the Nagios core
This change fixes a bug in the Nagios Core edition detection logic which could occur when using the
Nagios core with another edition than the Raw Edition. This is a rare use case, but can happen
temporarily during the migration from the Raw Edition to an Enterprise Edition.

The bug manifests itself by reporting the wrong edition via the livestatus interface. This can be
observed with the following query:

```
lq "GET status\nColumns: edition"
```

The edition column was introduced in 2.2.0.

Change-Id: I1d068a40e8624ff9c96d08b472e2496d7b5deaaa