Differentiate between old replacements and new replacements
Since the replacements contain the edition, an information derived from
the OMD version, the replacement macros now need to be treated
differently for computing the old version skel files and the new version
skel files.

Previously the same replacements were used to compute the content of
old and new version skel files, which lead to incorrect results during
updates.

So far, only the file etc/mk-livestatus/nagios.cfg is using the
macro "###EDITION###", so only this file is affected by that issue.
The result of the problem was that etc/mk-livestatus/nagios.cfg was not
rewritten with the new edition information when updating from a raw
edition to an enterprise edition.

Change-Id: I14fcbca9758634b9a8541bb1a0724424acbd2b5a