15722 FIX omd update: Fix setting world read permissions on etc and local files
Starting with 2.2.0 (see #15062), we use a umask of 0077 to prevent
files and directories owned by the site user from being world readable.
However, when updating from the CRE to a non CRE edition, some files and
directories were still set to be world readable during 'omd update'.
This is now in line with the other files.

Change-Id: I59561c44027bffc23dde045c3e19447cd7c8a09e