check levels: consistently create both metrics
With this we make the creation of the prediction
metrics more consistent:

The metric for lower levels is always called
'predict_lower_<name>', regardless of whether the
upper levels are predictive or not.

Both prediction metrics are recorded if present,
even if they are the same.
While it is nice to conslidate the output (for
human consumption) there is no benefit in not adding
the value to the RRD.

Change-Id: I7e73a05f51cbedf353e53b2a70a45053a6a1d921