15723 FIX Ensure site is stopped when running cmk-update-config
The `cmk-update-config` command, which is usually executed as part of `omd
update` expects the site to be stopped during execution. This is always ensured
during `omd update`.

However, the command can also be executed manually later on. In this situation
the site can be running. To prevent unexpected results, the command now protects
against an accidentally running site on invocation.

Change-Id: Ibaf565b5d6fac9e81da441f5d48353453baab1d6