16376 FIX DCD: Not respecting "Validity of missing data" setting
This fixes disappearing hosts in case of temporarily missing data around the time when
a cleanup background job is run (around midnight).

In case a piggybacked host temporarily did not receive any data while the background
(cron) job `cmk --cleanup-piggyback` was executed, a subsequent run of the DCD would not
respect the "Validity of missing data" setting, wrongly removing the affected host
from the monitoring configuration.

Change-Id: I97519bfab8bc76c21cedeafdede864fb51748df1