16330 FIX size_trend: avoid 'infinite to integer' crash
While calculating how much time is left until the resource is full, if the value of the mb in range was too low (e.g. 1e-320) the check used to crash with 'infinite to integer'.
This has now been fixed and if the value is too low and causes division with it to result in 'inf', the check will not be executed.

SUP-17029

Change-Id: I3093ceb20525e7945908a4fdacf5fd60f9fc9e1f