discovery: restrict changed services display to label change only
Prior to this commit, changed services in the 'Check_MK Discovery'
service included changes to both the discovered labels as well
as the parameters. However this led to the following confusions:

* changed services targeted both either labels or parameters, the user
  was unable to know which one in particular changed in the service
  summary
* the discovery page only showed changed services when labels changed
  but not the parameters as a front-end concept is yet to be made

For these reasons, we decided to remove the ability to modify the
discovered parameters as well as exclude them in the changed service
display. The removed functionality of this commit can be reintroduced
when a proper front-end concept on the discovery page has been decided.

Change-Id: Ib16ca004aa2a5b74874a0b9ff0d082c5a82b0bff