16048 Import Azure tags as service labels and provide regex matching
Azure resource tags are now imported as service labels as well (in addition to the existing host labels). Each service that is discovered for an Azure resource now holds the resource's tags as service labels.
Also, you can now restrict the import of Azure tags via regex or disable the import altogether. This can be done in the rule "Microsoft Azure" via option "Import tags as host/service labels".
Note that tags are validated to 1) not hold empty values - empty values are replaced by "true" and 2) not contain colons (":") - colons are replaced by underscores ("_"). So an Azure resource tag "tag:key":"" will now be shown as "tag_key":"true".

Change-Id: I4ae7c2034e79b1ae338c125c0aa7a993a0a45363