16465 FIX service discovery: introduce functionality to fetch job status from remote sites
Prior to this werk, the service discovery endpoints were restricted to the local
service discovery background job. This resulted in the following endpoints being
restricted to local setups only:

LI: Show the last service discovery background job on a host
LI: Wait for service discovery completion

This werk fixes this and now also supports distributed monitoring setups. To benefit
from this change both the central site as well the remote sites need to be updated
to the werk's version. This change does not affect local only setups.

Change-Id: Ib6820eb272be4096df821b1dc2c05d165ac9f54f