15321 FIX Fix "State if specific check plugins receive no monitoring data" of Rule "Status of the Checkmk service"
Rule "Status of the Checkmk service" provides a setting called "State if
specific check plugins receive no monitoring data" where you can specify a
regular expression to match specific check plugins, and assign a status for
the "Check_MK" service if this check plugins receives no data.

The feature did work correctly if you specified a Status worse than "WARN".

But the "Check_MK" service went to "WARN" even if there was an rule to set the
status to "OK" if the specific section did not receive any data. This is fixed now.

SUP-17271

Change-Id: I449cf23a3b9fce9227ff7d7f12a0162e8dc81c3e