16364 FIX tag_group: change the identification field from 'ident' to 'id'
Before this Werk, when creating a tag group, the **ident** field was used to provide its identification as well as that of the tags associated to the group, however the information was returned in the **id** field. This Werk unifies the names and now the **ident** fields are renamed to **id**. Users should adapt their scripts accordingly.

For comptatibility reasons CheckMK 2.2.0 and 2.3.0 will support both **id** and **ident**, but **ident** will be removed on the next version.

The following example shows the changes that need to be applied to the payload to use this endpoint:

Original payload:
```json
{
	"ident": "test_group",
	"title": "Test group",
	"help_text": "My test groupd",
	"tags": [
		{"ident": "test", "title": "Test Tag"}
	]
}
```

Updated payload:
```json
{
	"id": "test_group",
	"title": "Test group",
	"help_text": "My test groupd",
	"tags": [
		{"id": "test", "title": "Test Tag"}
	]
}
```

SUP-16902

Change-Id: I24fe59ea8ac1f6b30bc471c5e110d007d5334c71